update button functionality added

+++ src/pages/MyCart.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../providers/AuthProvider";
-import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import Swal from "sweetalert2";
 
 
 const MyCart = () => {
@@ -14,26 +14,43 @@ const MyCart = () => {
         console.log(data)});
     },[user.email]);
     const handleDelete = id => {
-        console.log(id);
-        fetch(`http://localhost:5010/carts/${id}`,{
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount){
-                toast('This product successfully deleted from your cart');
-                const remaining = cart.filter(aCart => aCart._id !== id);
-                setCart(remaining);
-            }
-        })
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+         
+
+          fetch(`http://localhost:5010/carts/${id}`,{
+              method: 'DELETE'
+          })
+          .then(res => res.json())
+          .then(data => {
+              if(data.deletedCount){
+                   Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+                  const remaining = cart.filter(aCart => aCart._id !== id);
+                  setCart(remaining);
+              }
+          })
+        }
+      })
     }
     if(loading){
         return <span className="loading loading-spinner loading-lg"></span>
     }
     return (
         <div>
-            {/* <h1 className="text-center text-7xl font-Raleway font-bold text-green-500">Cart</h1>
-            <hr /> */}
+            <h1 className="text-center text-7xl font-Raleway font-bold text-green-500">Cart</h1>
+            <hr />
            
            <div className="overflow-x-auto">
   <table className="table">
@@ -61,7 +78,6 @@ const MyCart = () => {
     </tbody>
   </table>
 </div>
-<ToastContainer />
         </div>
     );
 };
